Switch tests to mocha
